Refactored getBiography
Moved the getBiography function from the views folder to a new folder called biography in the api folder. Added a thoughts parameter to the function, changed the datastore.items to thoughts, and adjusted the prompt to follow the same naming convention. Wrote very basic comments, need to add more in depth documentation to the file, up to my level of understanding. -RB